requirements.txt
pytest==6.2.3
pytest-cov==2.11.1
rasa==2.4.3
requests==2.25.1
spacy==3.0.5
python-dotenv==0.17.0
discord==1.0.1
discord.py==1.7.0
